Moved mirror implementation to a separate file mf_ultralight_listener_i.c
